+++ pytestqt/qtbot.py
@@ -1,9 +1,10 @@
-import functools
 import contextlib
+import functools
 import weakref
 
-from pytestqt.wait_signal import SignalBlocker, MultiSignalBlocker, SignalTimeoutError, SignalEmittedSpy
 from pytestqt.qt_compat import qt_api
+from pytestqt.wait_signal import SignalBlocker, MultiSignalBlocker, SignalTimeoutError, SignalEmittedSpy, \
+    SignalEmittedError
 
 
 def _parse_ini_boolean(value):
@@ -26,6 +27,9 @@ class QtBot(object):
     **Widgets**
 
     .. automethod:: addWidget
+    .. automethod:: capture_exceptions
+    .. automethod:: waitActive
+    .. automethod:: waitExposed
     .. automethod:: waitForWindowShown
     .. automethod:: stopForInteraction
     .. automethod:: wait
@@ -141,6 +145,58 @@ class QtBot(object):
 
     add_widget = addWidget  # pep-8 alias
 
+    def waitActive(self, widget, timeout=1000):
+        """
+        Context manager that waits for timeout milliseconds or until the window is active.
+        If window is not exposed within ``timeout`` milliseconds, raise ``TimeoutError``.
+
+        This is mainly useful for asynchronous systems like X11, where a window will be mapped to screen
+        some time after  being asked to show itself on the screen.
+
+        .. code-block:: python
+
+            with qtbot.waitActive(widget, timeout=500):
+                show_action()
+
+        :param QWidget widget:
+            Widget to wait on.
+
+        :param int|None timeout:
+            How many miliseconds to wait on.
+
+        .. note::
+            This function is only available in PyQt5, raising a ``RuntimeError`` if called from
+            ``PyQt4`` or ``PySide``.
+        """
+        __tracebackhide__ = True
+        return _WaitWidgetContextManager('qWaitForWindowActive', 'activated', widget, timeout)
+
+    def waitExposed(self, widget, timeout=1000):
+        """
+        Context manager that waits for timeout milliseconds or until the window is exposed.
+        If the window is not exposed within ``timeout`` milliseconds, raise ``TimeoutError``.
+
+        This is mainly useful for asynchronous systems like X11, where a window will be mapped to screen
+        some time after  being asked to show itself on the screen.
+
+        .. code-block:: python
+
+            with qtbot.waitExposed(splash, timeout=500):
+                startup()
+
+        :param QWidget widget:
+            Widget to wait on.
+
+        :param int|None timeout:
+            How many miliseconds to wait on.
+
+        .. note::
+            This function is only available in PyQt5, raising a ``RuntimeError`` if called from
+            ``PyQt4`` or ``PySide``.
+        """
+        __tracebackhide__ = True
+        return _WaitWidgetContextManager('qWaitForWindowExposed', 'exposed', widget, timeout)
+
     def waitForWindowShown(self, widget):
         """
         Waits until the window is shown in the screen. This is mainly useful for asynchronous
@@ -150,17 +206,14 @@ class QtBot(object):
         :param QWidget widget:
             Widget to wait on.
 
-        .. note:: In Qt5, the actual method called is qWaitForWindowExposed,
-            but this name is kept for backward compatibility
+        .. note:: In ``PyQt5`` this function is considered deprecated in favor of :meth:`waitForWindowExposed`.
 
         .. note:: This method is also available as ``wait_for_window_shown`` (pep-8 alias)
         """
-        if hasattr(qt_api.QtTest.QTest, 'qWaitForWindowShown'):  # pragma: no cover
-            # PyQt4 and PySide
-            qt_api.QtTest.QTest.qWaitForWindowShown(widget)
-        else:  # pragma: no cover
-            # PyQt5
-            qt_api.QtTest.QTest.qWaitForWindowExposed(widget)
+        if qt_api.pytest_qt_api == 'pyqt5':
+            return qt_api.QtTest.QTest.qWaitForWindowExposed(widget)
+        else:
+            return qt_api.QtTest.QTest.qWaitForWindowShown(widget)
 
     wait_for_window_shown = waitForWindowShown  # pep-8 alias
 
@@ -390,7 +443,7 @@ class QtBot(object):
             qtbot.waitUntil(lambda: view_model.count() > 10)
 
         Note that this usage only accepts returning actual ``True`` and ``False`` values,
-        so returning an empty list to express "falseness" raises an ``ValueError``.
+        so returning an empty list to express "falseness" raises a ``ValueError``.
 
         :param callback: callable that will be called periodically.
         :param timeout: timeout value in ms.
@@ -495,8 +548,19 @@ class QtBot(object):
                 setattr(cls, method_name, method)
 
 
-# provide easy access to SignalTimeoutError to qtbot fixtures
+class TimeoutError(Exception):
+    """
+    .. versionadded:: 2.1
+
+    Exception thrown by :meth:`pytestqt.qtbot.QtBot.waitActive` and
+    :meth:`pytestqt.qtbot.QtBot.waitExposed` methods if a timeout occurs.
+    """
+    pass
+
+# provide easy access to exceptions to qtbot fixtures
 QtBot.SignalTimeoutError = SignalTimeoutError
+QtBot.SignalEmittedError = SignalEmittedError
+QtBot.TimeoutError = TimeoutError
 
 
 def _add_widget(item, widget):
@@ -527,3 +591,40 @@ def _iter_widgets(item):
     Iterates over widgets registered in the given pytest item.
     """
     return iter(getattr(item, 'qt_widgets', []))
+
+
+class _WaitWidgetContextManager(object):
+    """
+    Context manager implementation used by ``waitActive`` and ``waitExposed`` methods.
+    """
+
+    def __init__(self, method_name, adjective_name, widget, timeout):
+        """
+        :param str method_name: name ot the ``QtTest`` method to call to check if widget is active/exposed.
+        :param str adjective_name: "activated" or "exposed".
+        :param widget:
+        :param timeout:
+        """
+        self._method_name = method_name
+        self._adjective_name = adjective_name
+        self._widget = widget
+        self._timeout = timeout
+
+    def __enter__(self):
+        __tracebackhide__ = True
+        if qt_api.pytest_qt_api != 'pyqt5':
+            raise RuntimeError('Available in PyQt5 only')
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        __tracebackhide__ = True
+        try:
+            if exc_type is None:
+                method = getattr(qt_api.QtTest.QTest, self._method_name)
+                r = method(self._widget, self._timeout)
+                if not r:
+                    msg = 'widget {} not {} in {} ms.'.format(self._widget, self._adjective_name, self._timeout)
+                    raise TimeoutError(msg)
+        finally:
+            self._widget = None
+